+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # Change Log
 
 ## [Unreleased](https://github.com/dalance/svlint/compare/v0.1.0...Unreleased) - ReleaseDate
+
+* [Added] Error message about `.svlint.toml` is not found
+* [Fixed] Other file's lint results are shown